% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESTWGT.R
\docType{data}
\name{TESTWGT}
\alias{TESTWGT}
\title{A data set created by merging 1) "actual" data from a "gold standard" survey (A1, A2), and 2) data from another survey (Q1, Q2), including weights columns for that data (W1, W2). A1/Q1 and A2/Q2 are responses to the same two questions, asked to the same 10 respondents (ID), along the same 1-99 response scale.}
\format{A data frame with 10 rows and 7 variables
\describe{\item{ID, A1, A2, Q1, Q2, W1, W2}{Paired "actual"/survey data with weights columns for survey data}
}}
\source{
Example data generated by author
}
\usage{
TESTWGT
}
\description{
A data set created by merging 1) "actual" data from a "gold standard" survey (A1, A2), and 2) data from another survey (Q1, Q2), including weights columns for that data (W1, W2). A1/Q1 and A2/Q2 are responses to the same two questions, asked to the same 10 respondents (ID), along the same 1-99 response scale.
}
\keyword{datasets}
